requirements.txt
asn1crypto==1.4.0
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
base45==0.4.3
cbor2==5.4.2
certifi==2020.12.5
certvalidator==0.11.1
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cloudevents==1.2.0
coinmarketcap==5.0.3
constantly==15.1.0
container-app-conf==5.2.1
cose==0.9.dev8
cryptography==3.3.1
cycler==0.10.0
dateparser==1.0.0
deprecation==2.1.0
ecdsa==0.17.0
elasticsearch==7.10.1
Flask==2.0.2
flynn==1.0.0b2
functions-framework==2.3.0
gunicorn==20.1.0
html2text==2020.1.16
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.3.0
incremental==17.5.0
inflect==5.0.2
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.0
more-itertools==8.6.0
n26==3.2.1
numpy==1.20.1
oscrypto==1.2.1
packaging==21.0
pandas==1.2.3
Pillow==8.1.2
py-range-parse==1.0.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.9
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-binance==0.7.9
python-dateutil==2.8.1
python-dotenv==0.19.2
pytimeparse==1.1.8
pytz==2020.5
PyYAML==5.3.1
qrcode==7.3.1
regex==2020.11.13
requests==2.25.1
requests-cache==0.5.2
revolut==0.1.4
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
tabulate==0.8.7
tenacity==6.3.1
toml==0.10.2
Twisted==20.3.0
txaio==20.12.1
tzlocal==2.1
ujson==4.0.1
Unidecode==1.3.2
urllib3==1.26.2
voluptuous==0.12.1
watchdog==1.0.2
Werkzeug==2.0.2
zipp==3.4.0
zope.interface==5.2.0
